Rework directory structure to split project into two services
