+++ project/services/data_updater/scripts/fetch_trends.py
