+++ project/services/default/src/main/java/com/google/sps/data/BasicCompetitor.java
