+++ project/services/default/src/main/webapp/WEB-INF/web.xml
